Fix for an invalid 'to map threshold' value being passed in (cherry picked from commit 2814eee6fac2b1f5d7c17dbbc1a49273ff50a463)
Conflicts:

	kernel/src/main/java/org/neo4j/kernel/impl/transaction/ReadOnlyTxManager.java
	kernel/src/main/java/org/neo4j/kernel/impl/transaction/TxManager.java

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/batchinsert/RelationshipTypeHolder.java
@@ -26,9 +26,9 @@ import org.neo4j.kernel.impl.util.ArrayMap;
 public class RelationshipTypeHolder
 {
     private final ArrayMap<String,Integer> relTypes = 
-        new ArrayMap<String,Integer>( 5, false, false);
+        new ArrayMap<String,Integer>( (byte)5, false, false);
     private final ArrayMap<Integer,String> idToName = 
-        new ArrayMap<Integer,String>( 5, false, false);
+        new ArrayMap<Integer,String>( (byte)5, false, false);
     
     RelationshipTypeHolder( NameData[] types )
     {
