+++ community/kernel/src/main/java/org/neo4j/kernel/impl/core/PropertyIndexManager.java
@@ -40,12 +40,12 @@ import org.neo4j.kernel.impl.util.ArrayMap;
 public class PropertyIndexManager
 {
     private ArrayMap<String,List<PropertyIndex>> indexMap = 
-        new ArrayMap<String,List<PropertyIndex>>( 5, true, false );
+        new ArrayMap<String,List<PropertyIndex>>( (byte)5, true, false );
     private ArrayMap<Integer,PropertyIndex> idToIndexMap = 
-        new ArrayMap<Integer,PropertyIndex>( 9, true, false );
+        new ArrayMap<Integer,PropertyIndex>( (byte)9, true, false );
 
     private ArrayMap<Transaction,TxCommitHook> txCommitHooks = 
-        new ArrayMap<Transaction,TxCommitHook>( 5, true, false );
+        new ArrayMap<Transaction,TxCommitHook>( (byte)5, true, false );
 
     private final TransactionManager transactionManager;
     private final PersistenceManager persistenceManager;
@@ -63,8 +63,8 @@ public class PropertyIndexManager
 
     void clear()
     {
-        indexMap = new ArrayMap<String,List<PropertyIndex>>( 5, true, false );
-        idToIndexMap = new ArrayMap<Integer,PropertyIndex>( 9, true, false );
+        indexMap = new ArrayMap<String,List<PropertyIndex>>( (byte)5, true, false );
+        idToIndexMap = new ArrayMap<Integer,PropertyIndex>( (byte)9, true, false );
         txCommitHooks.clear();
     }
